+++ CHANGELOG.md
@@ -11,12 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - State variables are stored as int and float instead of int64 and float64 so they can be JSON serialized when saving
 - Bug that created spurious entries in ``inbound_order_pipeline``
 - Bug in which only order LT was respected in simulation if both order and shipment LTs are nonzero
+- Bug in default value for ``consistency_checks`` in ``simulation()``
 
 ### Changed
 - New format for storing and loading named instances, and better documentation
 - ``newsvendor.newsvendor_explicit()`` parameter renamed to ``revenue`` from ``selling_revenue`` for consistency
 
 ### Added
+- ``newsvendor.newsvendor_poisson_explicit()`` function added, for explicit-form newsvendor problem with Poisson demands
 - Various unit tests
 - More robust attribute handling for various objects